rmeove contracts v2
