+++ src/contracts/SettingsController.ts
@@ -3,7 +3,7 @@ import { contracts } from "constants/contracts";
 import Contract from "./Contract";
 
 export default class SettingsController extends Contract {
-  private static address = contracts.settingsController;
+  private static address = contracts["accounts/settings"];
 
   createEmbeddedUpdateEndowmentControllerMsg(
     payload: UpdateEndowmentControllerMsg
