+++ src/services/contract/custom/useAdminResourcesQuery.ts
@@ -5,7 +5,7 @@ import { queryContract } from "services/juno/queryContract";
 import { useGetWallet } from "contexts/WalletContext";
 import { idParamToNum } from "helpers";
 import { isJunoAddress } from "schemas/tests";
-import { contracts } from "constants/contractsV2";
+import { contracts } from "constants/contracts";
 import { adminRoutes, appRoutes } from "constants/routes";
 import { apCWs } from "./constants";
